+++ release-notes.md
@@ -3,7 +3,7 @@ Jupiter release notes
 
 ### 2017-07-11:
 - 升级Netty版本到4.1.13.Final
-- Spring.schema支持 JAcceptor配置 [\#29](https://github.com/fengjiachun/Jupiter/issues/29)
+- Spring.schema支持 JAcceptor/JConnector配置 [\#29](https://github.com/fengjiachun/Jupiter/issues/29)
 - Bug fix: 修复spring-support中<jupiter:property serializerType="xxx" />配置serializerType无效的问题 [\#27](https://github.com/fengjiachun/Jupiter/issues/27#event-1150209875)
 
 ### 2017-06-29: version 1.2.8